refactor: ws-server
